Updating GitHub to match library on www.
baseline-info

https://techwiki.juniper.net/Automation_Scripting/Junos_Automation_Library/baseline-info
Problem: The configuration in GitHub doesn't match the configuration
example in techwiki.
link-up

https://techwiki.juniper.net/Automation_Scripting/Junos_Automation_Library/link-up
Problem: The example output is empty on GitHub.
lo0-filter

https://techwiki.juniper.net/Automation_Scripting/Junos_Automation_Library/lo0-filter
Problem: This item is not in GitHub.
protect-config

https://techwiki.juniper.net/Automation_Scripting/Junos_Automation_Library/protect-config
Problem: This item is not in GitHub.
show-bgp-policy

https://techwiki.juniper.net/Automation_Scripting/Junos_Automation_Library/show-bgp-policy
Problem: The example configuration from the TechWiki page is not in
GitHub.
show-trunk

https://techwiki.juniper.net/Automation_Scripting/Junos_Automation_Library/show-trunk

https://github.com/Juniper/junoscriptorium/blob/c66cdd1be3ffa9135489c2a22aec2a4a47a142f7/library/juniper/op/display/show-trunk/example-1.conf
Problem: The example-1.conf file in GitHub is an example output and
should be renamed to example-1.output.
slax-doctor

https://techwiki.juniper.net/Automation_Scripting/Junos_Automation_Library/slax-doctor
Problem: The example Output from TechWiki isn't in GitHub.